requirements.txt
#datautil==0.4
Flask==0.12
Flask_JWT_Extended==1.3.0
Flask_SocketIO==2.9.4
neo4j_driver==1.5.2
pandas==0.22.0
passlib==1.7.0
setuptools==18.5
pytz==2013.7
requests==2.6.0
setuptools==18.5
simplekv==0.10.0
six==1.10.0
Werkzeug==0.11.15
xlrd==1.0.0
configobj==5.0.6
convertdate==2.1.1
python_dateutil==2.6.0
mock==2.0.0
nose_parameterized==0.6.0
ordereddict==1.1
pip==9.0.1
regex==2017.02.08
ruamel.base==1.0.0
typing==3.6.1
tzlocal==1.3
umalqurra==0.2
wheel==0.30.0a0
validate_email==1.3
pluginbase==0.5
python-magic==0.4.13
daterangeparser==1.3
elasticsearch-dsl==5.3
humanize==0.5.1
eventlet==0.22.1
